requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.5.0
Flask-Session==0.4.0
python-dotenv==0.19.0
mysql-connector-python==8.0.26
PyMySQL==1.1.0
Werkzeug==2.0.1
Jinja2==3.0.1
SQLAlchemy==1.4.23
alembic==1.7.1 